% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_hidden.r
\name{auth_hidden}
\alias{auth_hidden}
\title{Authenticate for the hidden/unofficial API}
\usage{
auth_hidden(cookiefile, live = interactive())
}
\arguments{
\item{cookiefile}{path to your cookiefile. Usually not needed after running
\link{auth_hidden} once. See \code{vignette("unofficial-api", package =
"traktok")} for more information on authentication.}

\item{live}{opens Chromium browser to guide you through the auth process
(experimental).}
}
\value{
nothing. Called to set up authentication
}
\description{
Guides you through authentication for the hidden/unofficial API.
To learn more, see the \href{https://jbgruber.github.io/traktok/articles/unofficial-api.html#authentication}{hidden API vignette}
or view it locally with \code{vignette("unofficial-api", package = "traktok")}.
}
\examples{
\dontrun{
# to run through the steps of authentication
auth_hidden()
# or point to a cookie file directly
auth_hidden("www.tiktok.com_cookies.txt")
}
}
