% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_research.r
\name{auth_research}
\alias{auth_research}
\title{Authenticate for the official research API}
\usage{
auth_research(client_key, client_secret)
}
\arguments{
\item{client_key}{Client key for authentication}

\item{client_secret}{Client secret for authentication}
}
\value{
An authentication token
}
\description{
Authenticate for the official research API
}
\details{
You need to apply for access to the API and get the key
and secret from TikTok. See
\url{https://developers.tiktok.com/products/research-api/} for more
information.
}
\examples{
\dontrun{
auth_research(client_key, client_secret)
}
}
