% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_research.r
\name{print.traktok_query}
\alias{print.traktok_query}
\title{Print a traktok query}
\usage{
\method{print}{traktok_query}(x, ...)
}
\arguments{
\item{x}{An object of class \code{traktok_query}}

\item{...}{Additional arguments passed to \code{lobstr::tree}}
}
\value{
nothing. Prints traktok query.
}
\description{
Print a traktok query as a tree
}
\examples{
query() |>
  query_and(field_name = "hashtag_name",
            operation = "EQ",
            field_values = "rstats") |>
  print()
}
