% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_research.r
\name{print.tt_results}
\alias{print.tt_results}
\title{Print search result}
\usage{
\method{print}{tt_results}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tt_results}}

\item{...}{not used.}
}
\description{
Print a traktok search results
}
