% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traktok_old.R
\name{tt_comments}
\alias{tt_comments}
\title{Get comments under a video.}
\usage{
tt_comments(video_urls, max_comments = Inf, sleep_pool = 1:10, ...)
}
\arguments{
\item{video_urls}{vector of URLs to TikTok videos.}

\item{max_comments}{max number of comments to return.}

\item{sleep_pool}{a vector of numbers from which a waiting period is randomly
drawn.}

\item{...}{handed to \link{tt_json}.}
}
\value{
a data.frame
}
\description{
Get comments under a video.
}
\details{
The function will wait between scraping two videos to make it less
  obvious that a scraper is accessing the site. The period is drawn randomly
  from the `sleep_pool` and multiplied by a random fraction.
}
\examples{
\dontrun{
tt_comments("https://www.tiktok.com/@tiktok/video/7106594312292453675")
}
}
