% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tt_get_cookies}
\alias{tt_get_cookies}
\title{Get cookies}
\usage{
tt_get_cookies(x, save = TRUE, name = "tiktok.com_cookies")
}
\arguments{
\item{x}{either file location, a cookie string, or can be left blank (see
Details).}

\item{save}{should a cookie file be saved in the default location. If x is
already the default cookie file, is ignored.}

\item{name}{name of the cookie file used for loading or saving a specific
cookie file.}
}
\value{
a named list of cookies
}
\description{
Wraps different way to get valid cookies for traktok. See
  details.
}
\details{
to make requests to TikTok, traktok needs user cookies. By running
  \code{tt_get_cookies()} without arguments, you can obtain anonymous
  cookies. Most functions will work with these. To get cookies for a logged
  in user, you can use a browser extension to export the necessary cookies
  from your browser (after visiting TikTok.com at least once). I can
  recommend "Get cookies.txt" for Chromium based browsers or "cookies.txt"
  for Firefox. Use \code{tt_get_cookies("tiktok.com_cookies.txt")} to save
  the cookies permanently. By default, cookies are stored in the location
  returned by \code{tools::R_user_dir("traktok", "config")}.
}
