% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_hidden.r
\name{tt_get_following_hidden}
\alias{tt_get_following_hidden}
\alias{tt_get_follower_hidden}
\title{Get followers and following of a user from the hidden API}
\usage{
tt_get_following_hidden(
  secuid,
  sleep_pool = 1:10,
  max_tries = 5L,
  cookiefile = NULL,
  verbose = TRUE
)

tt_get_follower_hidden(
  secuid,
  sleep_pool = 1:10,
  max_tries = 5L,
  cookiefile = NULL,
  verbose = TRUE
)
}
\arguments{
\item{secuid}{The secuid of a user. Currently no better way to obtain it than
opening a users site in the browser, right click to to view page source,
then search for "secUid".}

\item{sleep_pool}{a vector of numbers from which a waiting period is randomly
drawn.}

\item{max_tries}{how often to retry if a request fails.}

\item{cookiefile}{path to your cookiefile. Default is to request a new one
from TikTok.com and place it in the location returned by
\code{tools::R_user_dir("traktok", "config")} and set the option cookiefile
to this location.}

\item{verbose}{logical. Print status messages.}
}
\value{
a data.frame of followers
}
\description{
Get followers and following of a user from the hidden API
}
\examples{
\dontrun{
tt_get_follower_hidden("MS4wLjABAAAAun-1Cl5bjjlMXYCkKo58aPekMkPkkFkWB0y0-lSIJ-EMQ_1RLj1slviOVj0Vpuv9")
}
}
