% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_hidden.r
\name{tt_request_hidden}
\alias{tt_request_hidden}
\title{Get json string from a TikTok URL using the hidden API}
\usage{
tt_request_hidden(url, max_tries = 5L, cookiefile = NULL)
}
\arguments{
\item{url}{a URL to a TikTok video or account}

\item{max_tries}{how often should the request be tried before throwing an error}

\item{cookiefile}{path to your cookiefile. Default is to request a new one
from TikTok.com and place it in the location returned by
\code{tools::R_user_dir("traktok", "config")} and set the option cookiefile
to this location.}
}
\description{
Get json string from a TikTok URL using the hidden API
}
