% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traktok_old.R
\name{tt_search}
\alias{tt_search}
\title{Search videos}
\usage{
tt_search(...)
}
\arguments{
\item{...}{\code{search_hidden()}.}
}
\value{
a data.frame
}
\description{
This function was replaced by \code{search_hidden()}.
}
