% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traktok.R
\name{tt_search_hashtag}
\alias{tt_search_hashtag}
\title{Search videos using hashtag}
\usage{
tt_search_hashtag(
  hashtag,
  max_videos = Inf,
  cookiefile = getOption("cookiefile"),
  ...
)
}
\arguments{
\item{hashtag}{hashtag to query}

\item{max_videos}{max number of videos to return.}

\item{cookiefile}{path to your cookiefile. See details in \link{tt_json}.}

\item{...}{handed to \link{tt_json}.}
}
\value{
a data.frame
}
\description{
Search videos using hashtag
}
\examples{
\dontrun{
tt_search_hashtag("rstats", max_videos = 5L)
}
}
