% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_research.r
\name{tt_user_info_api}
\alias{tt_user_info_api}
\title{Lookup TikTok videos by a user using the research API}
\usage{
tt_user_info_api(username, fields = "all", verbose = TRUE, token = NULL)
}
\arguments{
\item{username}{name of the user to be queried}

\item{fields}{The fields to be returned (defaults to all)}

\item{verbose}{should the function print status updates to the
screen?}

\item{token}{The authentication token (usually supplied
automatically after running auth_research once)}
}
\value{
A data.frame of parsed TikTok videos the user has posted
}
\description{
Lookup TikTok videos by a user using the research API
}
\examples{
\dontrun{
tt_user_info_api("jbgruber")
}
}
