% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_hidden.r
\name{tt_user_info_hidden}
\alias{tt_user_info_hidden}
\title{Get infos about a user from the hidden API}
\usage{
tt_user_info_hidden(username, parse = TRUE)
}
\arguments{
\item{username}{A URL to a video or username.}

\item{parse}{Whether to parse the data into a data.frame (set to FALSE to get
the full list).}
}
\value{
A data.frame of user info.
}
\description{
Get infos about a user from the hidden API
}
\examples{
\dontrun{
df <- tt_user_info_hidden("https://www.tiktok.com/@fpoe_at")
}
}
