% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traktok.R
\name{tt_user_videos}
\alias{tt_user_videos}
\title{Get user videos}
\usage{
tt_user_videos(user_url, sleep_pool = 1:10)
}
\arguments{
\item{user_url}{vector of URLs to TikTok accounts.}

\item{sleep_pool}{a vector of numbers from which a waiting period is randomly
drawn.}
}
\value{
a data.frame
}
\description{
Get user videos
}
\details{
The function will wait between scraping two accounts to make it less
  obvious that a scraper is accessing the site. The period is drawn randomly
  from the `sleep_pool` and multiplied by a random fraction.
}
\examples{
\dontrun{
tt_user_videos("https://www.tiktok.com/@tiktok")
}
}
