% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traktok.R
\name{tt_videos}
\alias{tt_videos}
\title{Get video URLs and metadata from videos}
\usage{
tt_videos(
  video_urls,
  save_video = FALSE,
  overwrite = FALSE,
  dir = ".",
  cache_dir = NULL,
  sleep_pool = 1:10,
  ...
)
}
\arguments{
\item{video_urls}{vector of URLs to TikTok videos.}

\item{save_video}{logical. Should the videos be downloaded.}

\item{overwrite}{logical. If save_video=TRUE and the file already exists,
should it be overwritten?}

\item{dir}{directory to save videos files to.}

\item{cache_dir}{if set to a path, one RDS file with metadata will be written
to disk for each video. This is useful if you have many videos and want to
pick up where you left if something goes wrong.}

\item{sleep_pool}{a vector of numbers from which a waiting period is randomly
drawn.}

\item{...}{handed to \link{tt_json}.}
}
\value{
a data.frame
}
\description{
Get video URLs and metadata from videos
}
\details{
The function will wait between scraping two videos to make it less
  obvious that a scraper is accessing the site. The period is drawn randomly
  from the `sleep_pool` and multiplied by a random fraction.
}
\examples{
\dontrun{
tt_videos("https://www.tiktok.com/@tiktok/video/7106594312292453675")
}
}
