% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_hidden.r
\name{auth_hidden}
\alias{auth_hidden}
\title{Get cookies for the hidden API}
\usage{
auth_hidden(x = NULL, save = TRUE, name = "tiktok.com_cookies")
}
\arguments{
\item{x}{either file location, a cookie string, or can be left blank (see
Details).}

\item{save}{should a cookie file be saved in the default location. If x is
already the default cookie file, is ignored.}

\item{name}{name of the cookie file used for loading or saving a specific
cookie file.}
}
\value{
a named list of cookies
}
\description{
Wraps different ways to get valid cookies for traktok. See
  details.
}
\details{
to make requests to TikTok, traktok needs user cookies. By running
  \code{auth_hidden()} without arguments, you can obtain anonymous
  cookies. Most functions will not work with these, however. To get cookies
  for a logged in user, you can use a browser extension to export the
  necessary cookies from your browser (after visiting TikTok.com at least
  once). I can recommend "Get cookies.txt" for Chromium based browsers or
  "cookies.txt" for Firefox. See \code{vignette("authentication", package =
  "traktok")} for details. By default, cookies are stored in the location
  returned by \code{tools::R_user_dir("traktok", "config")}.
}
\examples{
\dontrun{
auth_hidden("www.tiktok.com_cookies.txt")
}
}
