% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_hidden.r
\name{auth_hidden}
\alias{auth_hidden}
\title{Authenticate for the hidden/unofficial API}
\usage{
auth_hidden(cookiefile)
}
\arguments{
\item{cookiefile}{optional path to your cookiefile. See
\code{vignette("unofficial-api", package = "traktok")} for more information
on authentication.}
}
\value{
nothing. Called to set up authentication
}
\description{
Guides you through authentication for the hidden/unofficial API#'
}
\examples{
\dontrun{
# to run through the steps of authentication
auth_hidden()
# or point to a cookie file directly
auth_hidden("www.tiktok.com_cookies.txt")
}
}
