% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_.r
\name{last_query}
\alias{last_query}
\title{Retrieve most recent query}
\usage{
last_query()
}
\value{
a list of unparsed videos
}
\description{
If \code{tt_query_videos} videos failes after already getting several pages,
you can use this function to get all videos that have been retrieved to far
from memory (does not work when the sesseion has crashed).
}
