% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traktok.R
\name{tt_json}
\alias{tt_json}
\title{Get json file from a TikTok URL}
\usage{
tt_json(url, cookiefile = getOption("cookiefile"))
}
\arguments{
\item{url}{a URL to a TikTok video or account}

\item{cookiefile}{path to your cookiefile. See details.}
}
\description{
Get json file from a TikTok URL
}
\details{
To get a valid cookiefile, you need to visit TikTok in your browser
  and then use, for example, the Browser extension "Get cookies.txt"
  (available for Chromium based Browsers an Firefox). If you experience
  errors, the cookies might have expired. Just open TikTok in your browser
  again and export a new file in this case.
}
