% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traktok_old.R
\name{tt_json}
\alias{tt_json}
\title{Get json file from a TikTok URL}
\usage{
tt_json(...)
}
\arguments{
\item{...}{\code{tt_request_hidden()}.}
}
\description{
This function was replaced by \code{tt_request_hidden()}.
}
