% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traktok.R
\name{tt_search}
\alias{tt_search}
\title{Search videos}
\usage{
tt_search(
  q,
  scope = "video",
  max_videos = Inf,
  offset = 0L,
  cookiefile = getOption("cookiefile"),
  ...
)
}
\arguments{
\item{q}{query as one string}

\item{scope}{either "video".}

\item{max_videos}{max number of videos to return.}

\item{offset}{how many videos to skip. For example, if you already have the
first X of a search.}

\item{cookiefile}{path to your cookiefile. Default is to request a new one
from TikTok.com and place it in the location returned by
\code{tools::R_user_dir("traktok", "config")} and set the option cookiefile
to this location.}

\item{...}{handed to \link{tt_json}.}
}
\value{
a data.frame
}
\description{
Search videos
}
\examples{
\dontrun{
tt_search("#rstats", scope = "video", max_videos = 5L)
}
}
